src/edu/waketech/ccave/test/BirdTest.java
////////////////////////////////////////////////////////////////////////////////
//  Course:   CSC 251 Spring 2018
//  Section:  CSC 251
// 
//  Project:  ColossalCave
//  File:     BirdTest.java
//  
//  Name:     Rebecca Fenter
//  Email:    rtfenter@my.waketech.edu
////////////////////////////////////////////////////////////////////////////////

/**
 * Testing all aspects of our class
 */
package edu.waketech.ccave.test;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertFalse;
import static org.junit.Assert.assertTrue;

import org.junit.Test;

import edu.waketech.ccave.item.CCaveItem;
import edu.waketech.ccave.item.Bird;

public class BirdTest {
	
	/**
	 * Testing long description
	 */

	@Test
	public void testGetLongDescription() {
		CCaveItem cci = new Bird();
		assertEquals("There is a bird. It could be useful to put in our cage. ", cci.getLongDescription());
	}
	
	/**
	 * Testing short description
	 */

	//@Test
	//public void testGetShortDescription() {
		//CCaveItem cci = new Bird();
		//assertEquals("Bird", cci.getShortDescription());
	//}
	
	/**
	 * Testing ID
	 */

	@Test
	public void testGetId() {
		CCaveItem cci = new Bird();
		assertEquals("bird", cci.getId());
	}
	
	/**
	 * Testing location
	 */

	@Test
	public void testGetLocation() {
		CCaveItem cci = new Bird();
		assertEquals("orangeriverchamber", cci.getLocation());
	}
	
	/**
	 * Testing ID- upper/lower case
	 */

	@Test
	public void testIsId0() {
		CCaveItem cci = new Bird();
		assertTrue(cci.isId("birD"));
	}
	
	/**
	 * Testing ID- false 
	 */

	@Test
	public void testIsId1() {
		CCaveItem cci = new Bird();
		assertFalse(cci.isId("sophie"));
	}

}



